Move the AlgorithmLookups test into the CSP-specific test suite.
